pyproject.toml
[tool.poetry]
name = "llama-samples"
version = "0.1.0"
description = ""
authors = ["Naohiro2g <587586+Naohiro2g@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-cpp-python = {url = "https://github.com/abetlen/llama-cpp-python/releases/download/v0.2.87/llama_cpp_python-0.2.87-cp311-cp311-win_amd64.whl"}
diskcache = "5.6.1"
jinja2 = "2.11.3"
markupsafe = "2.0"
numpy = "2.0.1"
typing-extensions = "4.6.1"
gradio = "^4.41.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
